Clean up junction-core modules
