+++ crates/junction-core/src/endpoints.rs
@@ -2,14 +2,6 @@ use junction_api::http::{RouteRetry, RouteTimeouts};
 use std::net::SocketAddr;
 use xds_api::pb::envoy::config::{core::v3 as xds_core, endpoint::v3 as xds_endpoint};
 
-// FIXME: do we need to to support returning filters/rewrites? does supporting
-//        header filters mean adding headers to Endpoint?
-//
-// FIXME: Vec<Endpoints> is probably the wrong thing to return from all our
-// resolve methods. We probably need a struct that has something like a list
-// of primary endpoints to cycle through on retries, and a seprate list of
-// endpoints to mirror traffic to. Figure that out once we support mirroring.
-
 /// An HTTP endpoint to make a request to.
 ///
 /// Endpoints contain both a target [url][crate::Url] that should be given to an
